requirements.txt
albumentations==1.3.1
llama-index
llama-index-core
llama-index-agent-openai
llama-index-retrievers-bm25
llama-index-vector-stores-qdrant
llama-index-embeddings-huggingface
llama-index-llms-gemini
llama-index-postprocessor-colbert-rerank
llama-index-embeddings-clip
nougat-ocr
pymupdf4llm
markdown-analysis
qdrant-client
openai-clip
rank_bm25
transformers
torch
python-dotenv
transformers==4.38.2